return the status of database to initial

+++ python/db/init_table.py
@@ -3,12 +3,12 @@ import sqlite3, os, sys
 db_file = './kotori.db'
 
 if not os.path.isfile(db_file):
-	print 'database file not exists...'
-	exit
+	print('database file not exists...')
+	exit()
 	
 conn = sqlite3.connect(db_file)
 	
-if sys.argv[1] == 'drop':
+if len(sys.argv) > 1 and sys.argv[1] == 'drop':
 	conn.execute('''DROP TABLE IF EXISTS image''')
 	conn.execute('''DROP TABLE IF EXISTS user''')
 	print('drop image and user table...')
